# api/swagger.yaml
basePath: /
definitions:
  model.AlternativeTitle:
    properties:
      english:
        type: string
      japanese:
        type: string
      synonym:
        type: string
    type: object
  model.Anime:
    properties:
      airingDate:
        $ref: '#/definitions/model.StartEndDate'
      alternativeTitles:
        $ref: '#/definitions/model.AlternativeTitle'
      broadcast:
        type: string
      duration:
        type: string
      episode:
        type: integer
      favorite:
        type: integer
      genres:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      id:
        type: integer
      image:
        type: string
      licensors:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      member:
        type: integer
      popularity:
        type: integer
      premiered:
        type: string
      producers:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      rank:
        type: integer
      rating:
        type: string
      related:
        $ref: '#/definitions/model.Related'
      score:
        type: number
      song:
        $ref: '#/definitions/model.Song'
      source:
        type: string
      status:
        type: string
      studios:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      synopsis:
        type: string
      title:
        type: string
      type:
        type: string
      video:
        type: string
      voter:
        type: integer
    type: object
...
  model.ArticleItem:
    properties:
      id:
        type: integer
      image:
        type: string
      is_advertorial:
        type: boolean
      is_spoiler:
        type: boolean
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      username:
        type: string
      view:
        type: integer
    type: object
...
  model.Character:
    properties:
      about:
        type: string
      favorite:
        type: integer
      id:
        type: integer
      image:
        type: string
      japaneseName:
        type: string
      name:
        type: string
      nickname:
        type: string
    type: object
  model.CharacterItem:
    properties:
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      role:
        type: string
      voiceActors:
        items:
          $ref: '#/definitions/model.Role'
        type: array
    type: object
...
  model.ClubItem:
    properties:
      id:
        type: integer
      member:
        type: integer
      name:
        type: string
    type: object
...
  model.Date:
    properties:
      day:
        type: integer
      month:
        type: integer
      year:
        type: integer
    type: object
  model.Episode:
    properties:
      airedDate:
        type: string
      episode:
        type: integer
      japaneseTitle:
        type: string
      link:
        type: string
      tag:
        type: string
      title:
        type: string
    type: object
  model.Item:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  model.NewsItem:
    properties:
      comment:
        type: integer
      content:
        type: string
      date:
        type: string
      forum_id:
        type: integer
      id:
        type: integer
      image:
        type: string
      title:
        type: string
      username:
        type: string
    type: object
...
  model.Recommendation:
    properties:
      recommended:
        $ref: '#/definitions/model.Source'
      source:
        $ref: '#/definitions/model.Source'
      users:
        items:
          $ref: '#/definitions/model.RecommendationUser'
        type: array
    type: object
  model.RecommendationUser:
    properties:
      content:
        type: string
      username:
        type: string
    type: object
  model.Related:
    properties:
      adaptation:
        items:
          $ref: '#/definitions/model.RelatedItem'
        type: array
      alternativeSetting:
        items:
          $ref: '#/definitions/model.RelatedItem'
        type: array
      alternativeVersion:
        items:
          $ref: '#/definitions/model.RelatedItem'
        type: array
      character:
        items:
          $ref: '#/definitions/model.RelatedItem'
        type: array
      fullStory:
        items:
          $ref: '#/definitions/model.RelatedItem'
        type: array
      other:
        items:
          $ref: '#/definitions/model.RelatedItem'
        type: array
      parentStory:
        items:
          $ref: '#/definitions/model.RelatedItem'
        type: array
      prequel:
        items:
          $ref: '#/definitions/model.RelatedItem'
        type: array
      sequel:
        items:
          $ref: '#/definitions/model.RelatedItem'
        type: array
      sideStory:
        items:
          $ref: '#/definitions/model.RelatedItem'
        type: array
      spinOff:
        items:
          $ref: '#/definitions/model.RelatedItem'
        type: array
      summary:
        items:
          $ref: '#/definitions/model.RelatedItem'
        type: array
    type: object
  model.RelatedItem:
    properties:
      id:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
  model.Review:
    properties:
      chapter:
        description: for manga
        type: string
      date:
        type: string
      episode:
        description: for anime
        type: string
      helpful:
        type: integer
      id:
        type: integer
      image:
        type: string
      review:
        type: string
      score:
        $ref: '#/definitions/model.ReviewScore'
      source:
        $ref: '#/definitions/model.Source'
      username:
        type: string
    type: object
  model.ReviewScore:
    properties:
      art:
        description: for both
        type: integer
      character:
        description: for both
        type: integer
      enjoyment:
        description: for both
        type: integer
      overall:
        description: for both
        type: integer
      sound:
        description: for anime only
        type: integer
      story:
        description: for both
        type: integer
    type: object
  model.Role:
    properties:
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  model.Song:
    properties:
      ending:
        items:
          type: string
        type: array
      opening:
        items:
          type: string
        type: array
    type: object
  model.Source:
    properties:
      id:
        type: integer
      image:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  model.StartEndDate:
    properties:
      end:
        $ref: '#/definitions/model.Date'
      start:
        $ref: '#/definitions/model.Date'
    type: object
  model.Stats:
    properties:
      score:
        $ref: '#/definitions/model.StatsScore'
      summary:
        $ref: '#/definitions/model.StatsSummary'
    type: object
  model.StatsScore:
    properties:
      "1":
        $ref: '#/definitions/model.StatsScoreItem'
      "2":
        $ref: '#/definitions/model.StatsScoreItem'
      "3":
        $ref: '#/definitions/model.StatsScoreItem'
      "4":
        $ref: '#/definitions/model.StatsScoreItem'
      "5":
        $ref: '#/definitions/model.StatsScoreItem'
      "6":
        $ref: '#/definitions/model.StatsScoreItem'
      "7":
        $ref: '#/definitions/model.StatsScoreItem'
      "8":
        $ref: '#/definitions/model.StatsScoreItem'
      "9":
        $ref: '#/definitions/model.StatsScoreItem'
      "10":
        $ref: '#/definitions/model.StatsScoreItem'
    type: object
  model.StatsScoreItem:
    properties:
      percent:
        type: number
      vote:
        type: integer
    type: object
  model.StatsSummary:
    properties:
      completed:
        type: integer
      current:
        type: integer
      dropped:
        type: integer
      onHold:
        type: integer
      planned:
        type: integer
      total:
        type: integer
    type: object
...
  model.Video:
    properties:
      episodes:
        items:
          $ref: '#/definitions/model.VideoEpisode'
        type: array
      promotions:
        items:
          $ref: '#/definitions/model.VideoPromo'
        type: array
    type: object
  model.VideoEpisode:
    properties:
      episode:
        type: integer
      link:
        type: string
      title:
        type: string
    type: object
  model.VideoPromo:
    properties:
      link:
        type: string
      title:
        type: string
    type: object
  utils.Response:
    properties:
      data:
        type: object
      message:
        type: string
      meta:
        type: object
      status:
        type: integer
    type: object
info:
  contact:
    email: axel.rl.404@gmail.com
    name: Axel
    url: https://github.com/rl404
  description: Scraping/parsing MyAnimeList website to a useful and easy-to-use data.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Malscraper API
paths:
  /anime/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Anime'
              type: object
      summary: Get anime details
      tags:
      - anime
  /anime/{id}/article:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ArticleItem'
                  type: array
              type: object
      summary: Get anime featured article list
      tags:
      - anime
  /anime/{id}/characters:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.CharacterItem'
                  type: array
              type: object
      summary: Get anime character list
      tags:
      - anime
  /anime/{id}/clubs:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ClubItem'
                  type: array
              type: object
      summary: Get anime club list
      tags:
      - anime
  /anime/{id}/episodes:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Episode'
                  type: array
              type: object
      summary: Get anime episode list
      tags:
      - anime
  /anime/{id}/more-info:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get anime more information
      tags:
      - anime
  /anime/{id}/news:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.NewsItem'
                  type: array
              type: object
      summary: Get anime news list
      tags:
      - anime
  /anime/{id}/pictures:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      summary: Get anime picture list
      tags:
      - anime
  /anime/{id}/recommendations:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Recommendation'
                  type: array
              type: object
      summary: Get anime recommendation list
      tags:
      - anime
  /anime/{id}/reviews:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Review'
                  type: array
              type: object
      summary: Get anime review list
      tags:
      - anime
  /anime/{id}/staff:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Role'
                  type: array
              type: object
      summary: Get anime staff list
      tags:
      - anime
  /anime/{id}/stats:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Stats'
              type: object
      summary: Get anime score stats
      tags:
      - anime
  /anime/{id}/videos:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Video'
              type: object
      summary: Get anime video list
      tags:
      - anime
...
  /character/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Character'
              type: object
      summary: Get character details
      tags:
      - character
  /character/{id}/anime:
    get:
      consumes:
      - application/json
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Role'
                  type: array
              type: object
      summary: Get character animeography list
      tags:
      - character
  /character/{id}/article:
    get:
      consumes:
      - application/json
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ArticleItem'
                  type: array
              type: object
      summary: Get character article featured list
      tags:
      - character
  /character/{id}/clubs:
    get:
      consumes:
      - application/json
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ClubItem'
                  type: array
              type: object
      summary: Get character club list
      tags:
      - character
  /character/{id}/manga:
    get:
      consumes:
      - application/json
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Role'
                  type: array
              type: object
      summary: Get character mangaography list
      tags:
      - character
  /character/{id}/pictures:
    get:
      consumes:
      - application/json
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      summary: Get character picture list
      tags:
      - character
...
schemes:
- http
- https
swagger: "2.0"